+++ packages/zowe-explorer/src/extending/ZoweExplorerApiRegister.ts
@@ -10,7 +10,7 @@
  */
 
 import * as vscode from "vscode";
-import * as zowex from "@zowe/zowex-for-zowe-explorer";
+import * as zowex from "zowex-for-zowe-explorer";
 import {
     DataSetAttributesProvider,
     IApiExplorerExtender,
@@ -130,7 +130,7 @@ export class ZoweExplorerApiRegister implements Types.IApiRegisterClient {
         this.registerJesApi(new ZoweExplorerZosmf.JesApi());
         this.registerCommandApi(new ZoweExplorerZosmf.CommandApi());
 
-        this.registerMvsApi(new zowex.SshMvsApi(this.getDataSetAttrProvider?.()));
+        this.registerMvsApi(new zowex.SshMvsApi());
         this.registerUssApi(new zowex.SshUssApi());
         this.registerJesApi(new zowex.SshJesApi());
         this.registerCommandApi(new zowex.SshCommandApi());