+++ packages/zowex-for-zowe-explorer/CHANGELOG.md
@@ -6,11 +6,18 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## Recent Changes
 
-- **Breaking:** Renamed contributed setting IDs from `zowe-native-proto` to `zowex`. All references to `zowe-native-proto` should be replaced with `zowex` in VS Code `settings.json` files. [#831](https://github.com/zowe/zowex/issues/831)
+- Updated the `promptForProfile` function to deprioritize project-level configurations and disable profile creation during zowex server uninstall and restart operations. [#4224](https://github.com/zowe/zowe-explorer-vscode/pull/4224)
 - **Breaking:** Refactored the code to be directly integrated into Zowe Explorer. [#4210](https://github.com/zowe/zowe-explorer-vscode/pull/4210)
 - **Breaking:** Changed the settings to align with Zowe Explorer. [#4210](https://github.com/zowe/zowe-explorer-vscode/pull/4210)
   - `zowex-vsce.requestTimeout` -> `zowe.settings.requestTimeout`
   - All others, replace the setting prefix `zowex-vsce` with `zowe.zowex`
+- Fixed error handling of methods in `SshMvsApi` class so that errors are thrown and propagated to extenders. [#917](https://github.com/zowe/zowex/issues/917)
+- Fixed error where the encoding was not being passed when attempting to open spool files with an encoding. [#983](https://github.com/zowe/zowex/issues/983)
+
+## `0.5.0`
+
+- **Breaking:** Renamed contributed setting IDs from `zowe-native-proto` to `zowex`. All references to `zowe-native-proto` should be replaced with `zowex` in VS Code `settings.json` files. [#831](https://github.com/zowe/zowex/issues/831)
+- Fixed an error that would overwrite a member that already existed when creating a member with the "Replace" option. [#642](https://github.com/zowe/zowex/issues/642)
 
 ## `0.4.0`